test(accordion): add tests for Accordion component
